Fix ChannelFactory on_submit_share_standard
For how ChannelFactory is implemented it need to know the gorup id of
the downstream that send a SubmitShare but this is not required by the
protocol. This commit add a map between every created channel id and the
group id to which it belong so that the group id can be easly retreived
by the channel id.